[#148072065] fixed api request for feedback

+++ src/extensions/nexus_integration/util/submitFeedback.ts
@@ -1,3 +1,5 @@
+import { log } from '../../../util/log';
+
 import * as fs from 'fs-extra-promise';
 import ZipT = require('node-7z');
 import request = require('request');
@@ -10,7 +12,10 @@ export class TimeoutError extends Error {
   }
 }
 
-function submitFeedback(APIKey: string, message: string, feedbackFiles: string[]): Promise<void> {
+function zipFiles(files: string[]): Promise<string> {
+  if (files.length === 0) {
+    return Promise.resolve(undefined);
+  }
   const Zip: typeof ZipT = require('node-7z');
   const task: ZipT = new Zip();
   return new Promise<string>((resolve, reject) => {
@@ -22,35 +27,53 @@ function submitFeedback(APIKey: string, message: string, feedbackFiles: string[]
       }
       return resolve(tmpPath);
     });
-  }).then(tmpPath => {
-    return task.add(tmpPath, feedbackFiles, { ssw: true })
-      .then(() => tmpPath);
-  }).then(tmpPath => new Promise<any>((resolve, reject) => {
-    const formData = {
-      message,
-      files: fs.createReadStream(tmpPath),
-    };
-    const headers = {
-      'content-type': 'application/json',
-    };
-    if (APIKey) {
-      headers['APIKEY'] = APIKey;
-    }
-    request.post({
-      headers,
-      url: 'https://api.nexusmods.com/v1/feedback/',
-      formData,
-      timeout: 15000,
-    }, (error, response, body) => {
-      fs.removeAsync(tmpPath)
-      .then(() => {
-        if (error !== null) {
-          return reject(error);
+  }).then(tmpPath =>
+    task.add(tmpPath, files, { ssw: true })
+      .then(() => tmpPath));
+}
+
+function submitFeedback(APIKey: string, message: string, feedbackFiles: string[]): Promise<void> {
+  return zipFiles(feedbackFiles)
+    .then(tmpPath => new Promise<any>((resolve, reject) => {
+      const formData = {
+        feedback_text: message,
+      };
+      if (tmpPath !== undefined) {
+        formData['feedback_file'] = fs.createReadStream(tmpPath);
+      }
+      const headers = {};
+
+      if (APIKey) {
+        headers['APIKEY'] = APIKey;
+      }
+
+      const url = APIKey === null
+        ? 'https://api.nexusmods.com/v1/feedbacks/anonymous'
+        : 'https://api.nexusmods.com/v1/feedbacks';
+      request.post({
+        headers,
+        url,
+        formData,
+        timeout: 15000,
+      }, (error, response, body) => {
+        // TODO: write out only the response once the api is done
+        log('debug', 'got response for feedback', { error, response, body });
+        if (tmpPath) {
+          fs.removeAsync(tmpPath)
+            .then(() => {
+              if (error !== null) {
+                return reject(error);
+              }
+              resolve();
+            });
+        } else {
+          if (error !== null) {
+            return reject(error);
+          }
+          resolve();
         }
-        resolve();
       });
-    });
-  }));
+    }));
 }
 
 export default submitFeedback;
